fixed applying unsigned short cast to offsets and scales

+++ GOES/GOES-17/GLM/src/glm_read.c
@@ -33,6 +33,27 @@
  * Read and unpack all the event data in the file. It will be loaded
  * into the pre-allocated array of struct event. 
  *
+ * Note from PUG-L2+ Vol. 5:
+ *
+ *   5.0.2 Unsigned Integer Processing
+ * 
+ *   The classic model for netCDF (used by the GS) does not support
+ *   unsigned integers larger than 8 bits. Many of the variables in
+ *   GOES-R netCDF files are unsigned integers of 16-bit or 32-bit
+ *   length. The following process is recommended to convert these
+ *   unsigned integers:
+ *
+ *   1. Retrieve the variable data from the netCDF file.
+ *
+ *   2. For this variable, retrieve the attribute “_Unsigned”.
+ *
+ *   3. If the “_Unsigned” attribute is set to “true” or “True”, then
+ *   cast the variable data to be unsigned.  The steps above must be
+ *   completed before applying the scale_factor and add_offset values
+ *   to convert from scaled integer to science units. Also, the
+ *   valid_range and _FillValue attribute values are to be governed by
+ *   the “_Unsigned” attribute.
+ *
  * @param ncid ID of already opened GLM file.
  * @param nevents The number of events.
  * @param event Pointer to already-allocated arrat of GLM_EVENT_T.
@@ -137,10 +158,14 @@ read_event_vars(int ncid, int nevents, GLM_EVENT_T *event)
     for (i = 0; i < nevents; i++)
     {
 	event[i].id = event_id[i];
-	event[i].time_offset = (float)((unsigned short)event_time_offset[i]) * event_time_offset_scale + event_time_offset_offset;
-	event[i].lat = (float)((unsigned short)event_lat[i]) * event_lat_scale + event_lat_offset;
-	event[i].lon = (float)((unsigned short)event_lon[i]) * event_lon_scale + event_lon_offset;
-	event[i].energy = (float)((unsigned short)event_energy[i]) * event_energy_scale + event_energy_offset;
+	event[i].time_offset = (float)((unsigned short)event_time_offset[i]) *
+	    (unsigned short)event_time_offset_scale + (unsigned short)event_time_offset_offset;
+	event[i].lat = (float)((unsigned short)event_lat[i]) *
+	    (unsigned short)event_lat_scale + (unsigned short)event_lat_offset;
+	event[i].lon = (float)((unsigned short)event_lon[i]) *
+	    (unsigned short)event_lon_scale + (unsigned short)event_lon_offset;
+	event[i].energy = (float)((unsigned short)event_energy[i]) *
+	    (unsigned short)event_energy_scale + (unsigned short)event_energy_offset;
 	event[i].parent_group_id = event_parent_group_id[i];
     }
 
@@ -302,11 +327,14 @@ read_group_vars(int ncid, int ngroups, GLM_GROUP_T *group)
     for (i = 0; i < ngroups; i++)
     {
 	group[i].id = group_id[i];
-	group[i].time_offset = (float)((unsigned short)group_time_offset[i]) * group_time_offset_scale + group_time_offset_offset;
+	group[i].time_offset = (float)((unsigned short)group_time_offset[i]) *
+	    (unsigned short)group_time_offset_scale + (unsigned short)group_time_offset_offset;
 	group[i].lat = group_lat[i];
 	group[i].lon = group_lon[i];
-	group[i].area = (float)((unsigned short)group_area[i]) * group_area_scale + group_area_offset;
-	group[i].energy = (float)((unsigned short)group_energy[i]) * group_energy_scale + group_energy_offset;
+	group[i].area = (float)((unsigned short)group_area[i]) *
+	    (unsigned short)group_area_scale + (unsigned short)group_area_offset;
+	group[i].energy = (float)((unsigned short)group_energy[i]) *
+	    (unsigned short)group_energy_scale + (unsigned short)group_energy_offset;
 	group[i].parent_flash_id = group_parent_flash_id[i];
 	group[i].quality_flag = group_quality_flag[i];
     }
@@ -488,17 +516,19 @@ read_flash_vars(int ncid, int nflashes, GLM_FLASH_T *flash)
     {
 	flash[i].id = flash_id[i];
 	flash[i].time_offset_of_first_event = (float)((unsigned short)flash_time_offset_of_first_event[i]) *
-	    flash_time_offset_of_first_event_scale + flash_time_offset_of_first_event_offset;
+	    (unsigned short)flash_time_offset_of_first_event_scale + (unsigned short)flash_time_offset_of_first_event_offset;
 	flash[i].time_offset_of_last_event = (float)((unsigned short)flash_time_offset_of_last_event[i]) *
-	    flash_time_offset_of_last_event_scale + flash_time_offset_of_last_event_offset;
+	    (unsigned short)flash_time_offset_of_last_event_scale + (unsigned short)flash_time_offset_of_last_event_offset;
 	flash[i].frame_time_offset_of_first_event = (float)((unsigned short)flash_frame_time_offset_of_first_event[i]) *
-	    flash_frame_time_offset_of_first_event_scale + flash_frame_time_offset_of_first_event_offset;
+	    (unsigned short)flash_frame_time_offset_of_first_event_scale + (unsigned short)flash_frame_time_offset_of_first_event_offset;
 	flash[i].frame_time_offset_of_last_event = (float)((unsigned short)flash_frame_time_offset_of_last_event[i]) *
-	    flash_frame_time_offset_of_last_event_scale + flash_frame_time_offset_of_last_event_offset;
+	    (unsigned short)flash_frame_time_offset_of_last_event_scale + (unsigned short)flash_frame_time_offset_of_last_event_offset;
 	flash[i].lat = flash_lat[i];
 	flash[i].lon = flash_lon[i];
-	flash[i].area = (float)((unsigned short)flash_area[i]) * flash_area_scale + flash_area_offset;
-	flash[i].energy = (float)((unsigned short)flash_energy[i]) * flash_energy_scale + flash_energy_offset;
+	flash[i].area = (float)((unsigned short)flash_area[i]) * (unsigned short)flash_area_scale +
+	    (unsigned short)flash_area_offset;
+	flash[i].energy = (float)((unsigned short)flash_energy[i]) * (unsigned short)flash_energy_scale +
+	    (unsigned short)flash_energy_offset;
 	flash[i].quality_flag = flash_quality_flag[i];
     }
